// src/lib/mealie/client/models/ShoppingListItemsCollectionOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ShoppingListItemOut_Output } from './ShoppingListItemOut_Output';
/**
 * Container for bulk shopping list item changes
 */
export type ShoppingListItemsCollectionOut = {
    createdItems?: Array<ShoppingListItemOut_Output>;
    updatedItems?: Array<ShoppingListItemOut_Output>;
    deletedItems?: Array<ShoppingListItemOut_Output>;
};